more

+++ data/experimentation/json_writer.py
@@ -1,5 +1,6 @@
 import os
 import gzip
+import glob
 import jsonlines
 
 DEFAULT_MAX_CHUNK_SIZE_BYTES = 1024 * 1024 * 1024 # 1 GB
@@ -8,18 +9,35 @@ class JsonWriter:
     def __init__(self, base_path, chunk_test_interval=10000, max_chunk_size_bytes=DEFAULT_MAX_CHUNK_SIZE_BYTES, verbose=False):
         self.base_path = base_path
 
-        self.chunk_num = 0
+        self.chunk_num = 0 
         self.num_writes = 0
 
         self.verbose = verbose
         self.chunk_test_interval = chunk_test_interval
         self.max_chunk_size_bytes = max_chunk_size_bytes
-
+        self.last_processed_doc = -1
+        self._last_chunk()
         self._open_chunk()
 
+    def _last_chunk(self):
+        """Finds the number of the last existing chunk."""
+        files = glob.glob(f'{self.base_path}*')
+        self.chunk_num = len(files)
+        if self.chunk_num > 0:
+            self.chunk_num -= 1  # Decrement to get the last existing chunk
+
     def _open_chunk(self):
-        self.file = gzip.open(self.temporary_path, 'wt')
-        self.writer = jsonlines.Writer(self.file)
+        # Check if the temporary file already exists and open it in append mode if it does
+        if os.path.exists(self.temporary_path):
+            with gzip.open(self.temporary_path, 'rb') as f:
+                self.num_writes = 0
+                for line in jsonlines.Reader(f):
+                    self.num_writes += 1 # Count existing lines
+                self.last_processed_doc = line.get("id")
+            self.file = gzip.open(self.temporary_path, 'at')  # Open in append mode
+        else:
+            self.file = gzip.open(self.temporary_path, 'wt')
+        self.writer = jsonlines.Writer(self.file, flush=True)
 
     def _log(self, msg):
         if self.verbose:
@@ -48,9 +66,11 @@ class JsonWriter:
                 self.chunk_num += 1
                 self._open_chunk()
 
-
-    def close(self):
+    def cleanup(self):
         self.file.flush()
         self.writer.close()
         self.file.close()
+
+    def close(self):
+        self.cleanup()
         os.rename(self.temporary_path, self.path)
